+++ application/usecases/get_statement_use_case.go
@@ -13,7 +13,7 @@ func NewGetStatementUseCase(repository repositories.ClientRepository) GetStateme
 	return GetStatementUseCase{repository}
 }
 
-func (g GetStatementUseCase) Execute(clientId int, clientLimit int) (model.Statement, error) {
+func (g GetStatementUseCase) Execute(clientId int, clientLimit int64) (model.Statement, error) {
 	statement, err := g.repository.GetStatement(clientId, clientLimit)
 	if err != nil {
 		return model.Statement{}, err
